refactor: adjustmenting css in other file
