+++ src/components/app-card/product-details/styles.js
@@ -1,3 +1,4 @@
+// vendors
 import styled from "@emotion/styled";
 import {
   Card as MantineCard,
