+++ src/components/app-card/products-image/styles.js
@@ -1,7 +1,10 @@
-import styled from "@emotion/styled";
+// vendors
 import { Card } from "@mantine/core";
 
-const StyledCard = styled(Card)`
+// style
+import styled from "@emotion/styled";
+
+export const StyledCard = styled(Card)`
   cursor: pointer;
   display: flex;
   margin-top: 13px;
@@ -11,7 +14,7 @@ const StyledCard = styled(Card)`
     theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.gray[1]};
 `;
 
-const StyledImageSection = styled(Card.Section)`
+export const StyledImageSection = styled(Card.Section)`
   display: flex;
   height: 200px;
   overflow: hidden;
@@ -20,19 +23,9 @@ const StyledImageSection = styled(Card.Section)`
     theme.colorScheme === "dark" ? theme.colors.dark[4] : theme.colors.gray[3]};
 `;
 
-const StyledImage = styled.img`
+export const StyledImage = styled.img`
   width: 100%;
   min-height: 270px;
   height: auto;
   object-fit: cover;
 `;
-
-export function ProductsImage({ image }) {
-  return (
-    <StyledCard withBorder p={0}>
-      <StyledImageSection>
-        <StyledImage src={image[0]} alt={image[0]} />
-      </StyledImageSection>
-    </StyledCard>
-  );
-}
