+++ src/components/create-account/authentication.jsx
@@ -1,3 +1,4 @@
+// vendors
 import {
   TextInput,
   PasswordInput,
