+++ src/components/create-account/fn-create-account.jsx
@@ -1,3 +1,4 @@
+// vendors
 import { Button, Container, FileInput, TextInput } from "@mantine/core";
 import { IconAt } from "@tabler/icons";
 import { PasswordInput } from "@mantine/core";
