+++ src/components/dark-theme.jsx
@@ -1,3 +1,4 @@
+// vendors
 import { ActionIcon, useMantineColorScheme } from "@mantine/core";
 import { IconSun, IconMoonStars } from "@tabler/icons";
