+++ src/pages/_document.jsx
@@ -1,3 +1,4 @@
+// vendors
 import { createGetInitialProps } from "@mantine/next";
 import Document, { Head, Html, Main, NextScript } from "next/document";
 
@@ -10,6 +11,7 @@ export default class _Document extends Document {
     return (
       <Html>
         <Head />
+        
         <body>
           <Main />
           <NextScript />
