+++ src/pages/api/product.js
@@ -1,3 +1,4 @@
+// api
 import { products } from "../../data/products";
 
 export function getProducts() {
