+++ src/pages/produtos/[category]/[productId]/index.jsx
@@ -1,8 +1,10 @@
+// vendors
 import { Container, Text } from "@mantine/core";
+
+// components
 import { FooterSocial } from "../../../../components/footer";
 import { HeaderSearch } from "../../../../components/header";
 import { getProduct, getProducts } from "../../../api/product";
-
 import { ProductDetails } from "../../../../components/app-card/product-details/product-details";
 
 const ProductId = ({ products, type, id, items, setItems }) => {
